+++ packages/reactabular-tree/CHANGELOG.md
@@ -4,6 +4,7 @@
   * Bug fix - Allow `tree.toggleChildren` to work without column `props` defined.
   * Feature - Add `tree.getImmediateChildren`.
   * Feature - Add `tree.moveRows`.
+  * Breaking - Refactor `tree.filter` as `({ fieldName, parentField = 'parent' }) => (rows) => filteredRows`.
 
 6.1.1 / 2016-10-27
 ==================